Correctly report 64-bit file offsets in the "Parsing mailbox from" message on systems with 32-bit long.
CVS-ID: mailbox.c 1.32

+++ mailbox.c
@@ -782,7 +782,7 @@ int mailbox_parse(const char *mailbox)
 
 	/* load messages into idx_message msgs[] */
 	if (!error) {
-		logtty("Parsing mailbox from %lu...\n", inc_ofs);
+		logtty("Parsing mailbox from %llu...\n", (unsigned long long)inc_ofs);
 		error = mailbox_parse_fd(fd);
 		inc_ofs = lseek(fd, 0, SEEK_CUR);
 		error |= unlock_fd(fd);
